+++ internal/generator/generator.go
@@ -5,8 +5,8 @@ import (
 	"path/filepath"
 
 	"github.com/Mad-Pixels/go-dyno/internal/generator/schema"
-	"github.com/Mad-Pixels/go-dyno/internal/templ"
 	"github.com/Mad-Pixels/go-dyno/internal/utils"
+	"github.com/Mad-Pixels/go-dyno/internal/utils/tmplkit"
 	v2 "github.com/Mad-Pixels/go-dyno/templates/v2"
 )
 
@@ -43,7 +43,7 @@ func Generate(config *Config) error {
 		SecondaryIndexes: spec.SecondaryIndexes(),
 	}
 
-	res := templ.MustParseTemplateFormattedToString(v2.CodeTemplate, schemaMap)
+	res := tmplkit.MustParseTemplateFormattedToString(v2.CodeTemplate, schemaMap)
 	if err := utils.WriteToFile(g, []byte(res)); err != nil {
 		return err
 	}
